// portal/ui/src/app/interfaces/ICommunication.ts
export interface IApi {
  organization: IOrganization;
  id: string;
  name: string;
  description: string;
  createdBy: string;
  image?: string;

  /** @format date-time */
  createdOn: string;

  /** @format int32 */
  numPublished: number;
}

export interface IApiGateway {
  gatewayId: string;
}

export interface IApiPlan {
  planId: string;
  version: string;
  discoverability: Discoverability;
}

export interface IApiVersion {
  /** @format int64 */
  id: number;
  api: IApi;
  status: 'Created' | 'Ready' | 'Published' | 'Retired';
  endpoint: string;
  endpointType: 'rest' | 'soap' | 'ui';
  endpointContentType: 'json' | 'xml';
  endpointProperties: Record<string, string>;
  gateways: IApiGateway[];
  publicAPI: boolean;
  plans: IApiPlan[];
  version: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
  modifiedBy: string;

  /** @format date-time */
  modifiedOn: string;

  /** @format date-time */
  publishedOn: string;

  /** @format date-time */
  retiredOn: string;
  definitionType:
    | 'None'
    | 'SwaggerJSON'
    | 'SwaggerYAML'
    | 'WSDL'
    | 'WADL'
    | 'RAML'
    | 'External';
  parsePayload: boolean;
  disableKeysStrip: boolean;
  definitionUrl: string;
  extendedDescription: string;
  publicDiscoverability: Discoverability;
}
export interface IOrganization {
  id: string;
  name: string;
  description: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
  modifiedBy: string;

  /** @format date-time */
  modifiedOn: string;
}

export interface IContractSummary {
  /** @format int64 */
  contractId: number;
  clientOrganizationId: string;
  clientOrganizationName: string;
  clientId: string;
  clientName: string;
  clientVersion: string;
  apiOrganizationId: string;
  apiOrganizationName: string;
  apiId: string;
  apiName: string;
  apiVersion: string;
  apiDescription: string;
  planName: string;
  planId: string;
  planVersion: string;

  /** @format date-time */
  createdOn: string;
}
// ...
export interface IClient {
  organization: IOrganization;
  id: string;
  name: string;
  description: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
}

export interface IClientVersion {
  /** @format int64 */
  id: number;
  client: IClient;
  status:
    | 'Created'
    | 'Ready'
    | 'Registered'
    | 'Retired'
    | 'AwaitingApproval'
    | 'Unregistered';
  version: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
  modifiedBy: string;

  /** @format date-time */
  modifiedOn: string;

  /** @format date-time */
  publishedOn: string;

  /** @format date-time */
  retiredOn: string;
  apikey: string;
}

export interface IContract {
  /** @format int64 */
  id: number;
  client: IClientVersion;
  api: IApiVersion;
  plan: IPlanVersion;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
}

export interface IPlan {
  organization: IOrganization;
  id: string;
  name: string;
  description: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
}

export interface IPlanVersion {
  /** @format int64 */
  id: number;
  plan: IPlan;
  status: 'Created' | 'Ready' | 'Locked';
  version: string;
  createdBy: string;

  /** @format date-time */
  createdOn: string;
  modifiedBy: string;

  /** @format date-time */
  modifiedOn: string;

  /** @format date-time */
  lockedOn: string;
}
// ...
export enum Discoverability {
  PORTAL = 'PORTAL',
  ANONYMOUS = 'ANONYMOUS',
  FULL_PLATFORM_MEMBERS = 'FULL_PLATFORM_MEMBERS',
  ORG_MEMBERS = 'ORG_MEMBERS'
}